% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregFigAndeler.R
\name{hisregFigAndeler}
\alias{hisregFigAndeler}
\title{Lag søylediagram eller som viser andeler av ulike variabler}
\usage{
hisregFigAndeler(RegData, valgtVar, datoFra = "2000-01-01",
  datoTil = "2050-01-01", reshID, minald = 0, maxald = 120, erMann = 99,
  outfile = "", forlop1 = 99, forlop2 = 99, enhetsUtvalg = 1,
  preprosess = F, hentData = F)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{forlop1}{Hovedforløpstype
1: Kirurgisk intervensjon
2: Medisinsk intervensjon
3: Kirurgisk og medisinsk intervensjon
4: Ingen intervensjon}

\item{forlop2}{Underforløp
1 - Eksisjon lukket med sutur
2 - Eksisjon med åpen granulering og sekunder tilheling
3 - Eksisjon med påfølgende hudplantasjon
4 - CO2-laser
5 - Deroofing}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{preprosess}{Preprosesser data
FALSE: Nei (Default)
TRUE: Ja}

\item{hentData}{Gjør spørring mot database
FALSE: Nei, RegData gis som input til funksjonen (Default)
TRUE: Ja}
}
\value{
En figur med søylediagram av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram som viser andeler av valgt variabel
filtrert på de utvalg som er gjort.
}

