% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregForlPasTabell.R
\name{hisregForlPasTabell}
\alias{hisregForlPasTabell}
\title{Forløps of pasient tabeller for hisreg}
\usage{
hisregForlPasTabell(
  RegDt,
  tidFra = "2008-01-01",
  tidTil = Sys.Date(),
  aldmin = 0,
  aldmax = 120,
  kjoen = 99,
  tidenh = "aar",
  frlType = 1,
  IDType = "PasientID"
)
}
\arguments{
\item{RegDt}{en dataramme med alle variabler}

\item{aldmin}{alder fra.}

\item{aldmax}{alder til}

\item{kjoen}{kjønn der Kvinne = 0
Mann = 1
begge = 99}

\item{tidenh}{tidsenhet. kan være "aar" eller "maaned"}

\item{frlType}{forløpstype
1: Kirurgisk intervensjon
2: Medisinsk intervensjon
3: Kirurgisk og medisinsk intervensjon
4: Ingen intervensjon}

\item{tidfra}{start dato. format: "YYYY-mm-dd"}

\item{tidtil}{slutt dato.  format: "YYYY-mm-dd"}

\item{IDtype}{ID for filtrering. "m_mceid" eller "PasientID"}
}
\value{
tabelldata med antall pasienter/forløp per sykhus
per tidsenhet
}
\description{
Denne funksjonen tar inn registerdata og lager tabeller
av antall pasienter/forløp per sykehus per tidsenhet
}
