% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregHentData.R
\name{hisregHentRegData}
\alias{hisregHentRegData}
\title{Provide global dataframe for NoRGast}
\usage{
hisregHentRegData()
}
\value{
RegData data frame
}
\description{
Provides Hisreg data from staging
}

