% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregKonverterData.R
\name{hisregKonverterData}
\alias{hisregKonverterData}
\title{Tilpass gammelt datasett til nytt}
\usage{
hisregKonverterData(RegData, ForlopsData, Followups)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\value{
RegData En dataramme med konvertert datasett
}
\description{
Denne funksjonen konverterer navn og verdier i gamle Hisreg til navn og verdier i det nye
}
