% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregPreprosess.R
\name{hisregPreprosess}
\alias{hisregPreprosess}
\title{Preprosesser hisregdata}
\usage{
hisregPreprosess(RegData)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}
}
\description{
Denne funksjonen gjør en del nøvendig preprosessering for at Hisreg sitt datasett
kan brukes i rapporter
}

