% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregSkjemaTabell.R
\name{hisregSkjemaTabell}
\alias{hisregSkjemaTabell}
\title{Skjematabeller for hisreg}
\usage{
hisregSkjemaTabell(
  df,
  tidFra = "2008-01-01",
  tidTil = Sys.Date(),
  status = 1,
  typeDato = "HovedDato"
)
}
\arguments{
\item{df}{en dataramme med alle variabler}

\item{status}{skjema status
1: Ferdigstilt
2: Ikke ferdig}

\item{typeDato}{Dato for filtrering. "HovedDato"
"OpprettetDato"
"SistLagtretDato"}

\item{tidfra}{start dato. format: "YYYY-mm-dd"}

\item{tidtil}{slutt dato.  format: "YYYY-mm-dd"}
}
\value{
tabelldata med antall ferdigstilte/ikke ferdige
skjema per sykehus
}
\description{
Denne funksjonen tar inn skjemaoversikt og lager tabeller
av antall skjematyper per sykehus
}
