% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregUtvalg.R
\name{hisregUtvalg}
\alias{hisregUtvalg}
\title{Gjør utvalg av dataene}
\usage{
hisregUtvalg(RegData, datoFra, datoTil, minald, maxald, erMann, forlop1,
  forlop2, fargepalett = "BlaaRapp")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren). Tekststreng skrevet som 'YYYY-mm-dd'}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (Default)}

\item{forlop1}{Hovedforløpstype
1: Kirurgisk intervensjon
2: Medisinsk intervensjon
3: Kirurgisk og medisinsk intervensjon
4: Ingen intervensjon}

\item{forlop2}{Underforløp
1 - Eksisjon lukket med sutur
2 - Eksisjon med åpen granulering og sekunder tilheling
3 - Eksisjon med påfølgende hudplantasjon
4 - CO2-laser
5 - Deroofing}

\item{fargepalett}{Hvilken fargepalett skal brukes i figurer (Default: BlaaRapp)}
}
\value{
UtData En liste bestående av det filtrerte datasettet, utvalgstekst for figur og tekststreng som angir fargepalett
}
\description{
Denne funksjonen gjør utvalg av dataene og returnerer det filtrerte datasettet, utvalgsteksten
og fargepaletten for bruk i figuren
}

