% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastShinyHisreg.R
\name{lastShinyHisreg}
\alias{lastShinyHisreg}
\title{Last og tilrettelegg shiny-data}
\usage{
lastShinyHisreg()
}
\value{
En liste med data til bruk i hisreg sin shiny-app
}
\description{
Hvis det eksiterer, hent ferdig koblet og tilrettelagt data fra staging. Ellers
hent fra db og gjør nødvendige koblinger/mappinger.
}
