% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hisregHentData.R
\name{hisregHentTabell}
\alias{hisregHentTabell}
\title{Fetch chosen table for Hisreg}
\usage{
hisregHentTabell(tabnavn, registryName = "hisreg2")
}
\arguments{
\item{tabnavn}{Name of requested table}
}
\value{
RegData data frame
}
\description{
Provides Hisreg data from staging
}
